refactor: Moved contribution guidelines to a new document and changed name of page
